Make dir unit tests pass on Windows
* Use the filepath package instead of the path package
* Don't compile the dir.Init test on Windows; umasks aren't a thing
  there
