+++ internal/pkg/dir/dir_posix_test.go
@@ -1,0 +1,54 @@
+//go:build !windows
+// +build !windows
+
+/*
+Copyright 2022 k0s authors
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package dir_test
+
+import (
+	"os"
+	"path"
+	"path/filepath"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+	"golang.org/x/sys/unix"
+)
+
+func TestInit(t *testing.T) {
+	dir := t.TempDir()
+
+	foo := filepath.Join(dir, "foo")
+	require.NoError(t, dir.Init(foo, 0700), "failed to create temp dir foo")
+
+	checkPermissions(t, foo, 0700)
+
+	oldUmask := unix.Umask(0027)
+	t.Cleanup(func() { unix.Umask(oldUmask) })
+
+	bar := filepath.Join(dir, "bar")
+	require.NoError(t, Init(bar, 0755), "failed to create temp dir bar")
+	checkPermissions(t, bar, 0755)
+}
+
+// CheckPermissions checks the correct permissions
+func checkPermissions(t *testing.T, path string, want os.FileMode) {
+	info, err := os.Stat(path)
+	require.NoError(t, err, path)
+	assert.Equalf(t, want, info.Mode().Perm(), "%s has unexpected permissions", path)
+}
